+++ src/iguana/algorithms/clas12/ZVertexFilter/Validator.cc
@@ -60,7 +60,7 @@ namespace iguana::clas12 {
     for(auto const& row : particle_bank.getRowList()) {
       double vz = particle_bank.getFloat("vz", row);
       int pdg = particle_bank.getInt("pid", row);
-      int status = particle_bank.getInt("status", row);
+      int status = particle_bank.getShort("status", row);
       auto it = u_zvertexplots.find(pdg);
       //check if pdg is amongs those that we want to plot
       if (it != u_zvertexplots.end() && abs(status)>=2000) {
@@ -75,7 +75,7 @@ namespace iguana::clas12 {
     for(auto const& row : particle_bank.getRowList()) {
       double vz = particle_bank.getFloat("vz", row);
       int pdg = particle_bank.getInt("pid", row);
-      int status = particle_bank.getInt("status", row);
+      int status = particle_bank.getShort("status", row);
       auto it = u_zvertexplots.find(pdg);
       //check if pdg is amongs those that we want to plot
       if (it != u_zvertexplots.end() && abs(status)>=2000) {
